+++ lib/hermes/mcp/message.ex
@@ -1,120 +1,175 @@
 defmodule Hermes.MCP.Message do
   @moduledoc """
-  Handles encoding and decoding of MCP protocol messages.
+  Handles parsing and validation of MCP (Model Context Protocol) messages using the Peri library.
 
-  This module provides functions for parsing, validating, and creating
-  MCP (Model Context Protocol) messages using the JSON-RPC 2.0 format.
-
-  ## Message Types
-
-  The MCP protocol uses JSON-RPC 2.0 messages in these forms:
-
-  1. Requests - Messages sent from client to server that expect a response
-  2. Responses - Messages sent from server to client in response to a request
-  3. Notifications - Messages sent in either direction that don't expect a response
-  4. Errors - Error responses sent from server to client
-
-  ## Examples
-
-  ```elixir
-  # Decode MCP messages from a string
-  {:ok, messages} = Hermes.MCP.Message.decode(json_string)
-
-  # Encode a request
-  {:ok, request_json} = Hermes.MCP.Message.encode_request(%{"method" => "ping", "params" => %{}}, "req_123")
-
-  # Encode a notification
-  {:ok, notif_json} = Hermes.MCP.Message.encode_notification(%{"method" => "notifications/progress", "params" => %{...}})
-  ```
+  This module provides functions to parse and validate MCP messages based on the Model Context Protocol schema
   """
 
-  alias Hermes.MCP.Error
-  alias Hermes.MCP.Response
+  import Peri
+
+  # MCP message schemas
 
   @request_methods ~w(initialize ping resources/list resources/read prompts/get prompts/list tools/call tools/list logging/setLevel)
 
-  @notification_methods ~w(notifications/initialized notifications/cancelled notifications/progress notifications/message)
+  @init_params_schema %{
+    "protocolVersion" => {:required, :string},
+    "capabilities" => {:required, :map},
+    "clientInfo" => %{
+      "name" => {:required, :string},
+      "version" => {:required, :string}
+    }
+  }
 
-  @typedoc """
-  Represents any MCP protocol message.
+  @ping_params_schema :map
 
-  MCP protocol uses JSON-RPC 2.0 with several message formats:
+  @resources_list_params_schema %{
+    "cursor" => :string
+  }
 
-  1. Request Message:
-     ```
-     {
-       "jsonrpc": "2.0",
-       "method": String, // Method name like "ping", "resources/list"
-       "params": Object, // Method parameters
-       "id": String|Number // Request identifier
-     }
-     ```
+  @resources_read_params_schema %{
+    "uri" => {:required, :string}
+  }
 
-  2. Notification Message:
-     ```
-     {
-       "jsonrpc": "2.0",
-       "method": String, // Method name like "notifications/progress"
-       "params": Object  // Notification parameters
-     }
-     ```
+  @prompts_list_params_schema %{
+    "cursor" => :string
+  }
 
-  3. Response Message:
-     ```
-     {
-       "jsonrpc": "2.0",
-       "result": Object, // Result value
-       "id": String|Number // Request identifier
-     }
-     ```
+  @prompts_get_params_schema %{
+    "name" => {:required, :string},
+    "arguments" => :map
+  }
 
-  4. Error Message:
-     ```
-     {
-       "jsonrpc": "2.0",
-       "error": {
-         "code": Number,    // Error code
-         "message": String, // Error message
-         "data": Any        // Optional error data
-       },
-       "id": String|Number // Request identifier
-     }
-     ```
+  @tools_list_params_schema %{
+    "cursor" => :string
+  }
+
+  @tools_call_params_schema %{
+    "name" => {:required, :string},
+    "arguments" => :map
+  }
+
+  @log_levels ~w(debug info notice warning error critical alert emergency)
+
+  @set_log_level_params_schema %{
+    "level" => {:required, {:enum, @log_levels}}
+  }
+
+  defschema :request_schema, %{
+    "jsonrpc" => {:required, {:string, {:eq, "2.0"}}},
+    "method" => {:required, {:enum, @request_methods}},
+    "params" => {:dependent, &parse_request_params_by_method/1},
+    "id" => {:required, {:either, {:string, :integer}}}
+  }
+
+  defp parse_request_params_by_method(%{"method" => "initialize"}), do: {:ok, @init_params_schema}
+  defp parse_request_params_by_method(%{"method" => "ping"}), do: {:ok, @ping_params_schema}
+
+  defp parse_request_params_by_method(%{"method" => "resources/list"}), do: {:ok, @resources_list_params_schema}
+
+  defp parse_request_params_by_method(%{"method" => "resources/read"}), do: {:ok, @resources_read_params_schema}
+
+  defp parse_request_params_by_method(%{"method" => "prompts/list"}), do: {:ok, @prompts_list_params_schema}
+
+  defp parse_request_params_by_method(%{"method" => "prompts/get"}), do: {:ok, @prompts_get_params_schema}
+
+  defp parse_request_params_by_method(%{"method" => "tools/list"}), do: {:ok, @tools_list_params_schema}
+
+  defp parse_request_params_by_method(%{"method" => "tools/call"}), do: {:ok, @tools_call_params_schema}
+
+  defp parse_request_params_by_method(%{"method" => "logging/setLevel"}), do: {:ok, @set_log_level_params_schema}
+
+  defp parse_request_params_by_method(_), do: {:ok, :map}
+
+  @init_noti_params_schema :map
+  @cancel_noti_params_schema %{
+    "requestId" => {:required, {:either, {:string, :integer}}},
+    "reason" => :string
+  }
+  @progress_notif_params_schema %{
+    "progressToken" => {:required, {:either, {:string, :integer}}},
+    "progress" => {:required, {:either, {:float, :integer}}},
+    "total" => {:either, {:float, :integer}}
+  }
+  @logging_message_notif_params_schema %{
+    "level" => {:required, {:enum, @log_levels}},
+    "data" => {:required, :any},
+    "logger" => :string
+  }
+
+  defschema :notification_schema, %{
+    "jsonrpc" => {:required, {:string, {:eq, "2.0"}}},
+    "method" =>
+      {:required,
+       {:enum, ~w(notifications/initialized notifications/cancelled notifications/progress notifications/message)}},
+    "params" => {:dependent, &parse_notification_params_by_method/1}
+  }
+
+  defp parse_notification_params_by_method(%{"method" => "notifications/initialized"}),
+    do: {:ok, @init_noti_params_schema}
+
+  defp parse_notification_params_by_method(%{"method" => "notifications/cancelled"}),
+    do: {:ok, @cancel_noti_params_schema}
+
+  defp parse_notification_params_by_method(%{"method" => "notifications/progress"}),
+    do: {:ok, @progress_notif_params_schema}
+
+  defp parse_notification_params_by_method(%{"method" => "notifications/message"}),
+    do: {:ok, @logging_message_notif_params_schema}
+
+  defp parse_notification_params_by_method(_), do: {:ok, :map}
+
+  defschema :response_schema, %{
+    "jsonrpc" => {:required, {:string, {:eq, "2.0"}}},
+    "result" => {:required, :any},
+    "id" => {:required, {:either, {:string, :integer}}}
+  }
+
+  defschema :error_schema, %{
+    "jsonrpc" => {:required, {:string, {:eq, "2.0"}}},
+    "error" => %{
+      "code" => {:required, :integer},
+      "message" => {:required, :string},
+      "data" => :any
+    },
+    "id" => {:required, {:either, {:string, :integer}}}
+  }
+
+  defschema :mcp_message_schema,
+            {:oneof,
+             [
+               get_schema(:request_schema),
+               get_schema(:notification_schema),
+               get_schema(:response_schema),
+               get_schema(:error_schema)
+             ]}
+
+  @doc """
+  Determines if a JSON-RPC message is a request.
   """
-  @type message :: map()
-
-  # Message type guard functions
   defguard is_request(data) when is_map_key(data, "method") and is_map_key(data, "id")
+
+  @doc """
+  Determines if a JSON-RPC message is a notification.
+  """
   defguard is_notification(data) when is_map_key(data, "method") and not is_map_key(data, "id")
+
+  @doc """
+  Determines if a JSON-RPC message is a response.
+  """
   defguard is_response(data) when is_map_key(data, "result") and is_map_key(data, "id")
+
+  @doc """
+  Determines if a JSON-RPC message is an error.
+  """
   defguard is_error(data) when is_map_key(data, "error") and is_map_key(data, "id")
 
   @doc """
-  Decodes a JSON string into MCP message(s).
+  Decodes raw data (possibly containing multiple messages) into JSON-RPC messages.
 
-  This function handles both single messages and newline-delimited message streams.
-
-  ## Parameters
-
-    * `data` - The JSON string to decode
-
-  ## Returns
-
-    * `{:ok, messages}` where messages is a list of parsed MCP messages
-    * `{:error, error}` if parsing fails
-
-  ## Examples
-
-      iex> Hermes.MCP.Message.decode(~s({"jsonrpc":"2.0","result":{},"id":"req_123"}))
-      {:ok, [%{"jsonrpc" => "2.0", "result" => %{}, "id" => "req_123"}]}
-      
-      iex> {:error, error} = Hermes.MCP.Message.decode("invalid")
-      iex> error.code
-      -32700
-      iex> error.reason
-      :parse_error
+  Returns either:
+  - `{:ok, messages}` where messages is a list of parsed JSON-RPC messages
+  - `{:error, reason}` if parsing fails
   """
-  @spec decode(String.t()) :: {:ok, list(message())} | {:error, Error.t()}
   def decode(data) when is_binary(data) do
     data
     |> String.split("\n", trim: true)
@@ -126,81 +181,52 @@ defmodule Hermes.MCP.Message do
   end
 
   defp parse_message(line, {:ok, acc}) do
-    case JSON.decode(line) do
-      {:ok, message} ->
-        {:cont, {:ok, [message | acc]}}
-
-      {:error, _} ->
-        {:halt, {:error, Error.parse_error(%{line: line})}}
+    with {:ok, message} <- JSON.decode(line),
+         {:ok, message} <- validate_message(message) do
+      {:cont, {:ok, [message | acc]}}
+    else
+      err -> {:halt, err}
     end
   end
 
   @doc """
-  Encodes a request message into a JSON-RPC 2.0 compliant string.
-
-  ## Parameters
-
-    * `request` - A map containing the request method and params
-    * `id` - A unique identifier for the request
-
-  ## Returns
-
-    * `{:ok, json_string}` with the encoded request and a newline
-    * `{:error, error}` if encoding fails
-
-  ## Examples
-
-      iex> {:ok, request} = Hermes.MCP.Message.encode_request(%{"method" => "ping", "params" => %{}}, "req_123")
-      iex> request_map = request |> String.trim() |> JSON.decode!()
-      iex> request_map["jsonrpc"]
-      "2.0"
-      iex> request_map["method"]
-      "ping"
-      iex> request_map["id"]
-      "req_123"
+  Validates a decoded JSON message to ensure it complies with the MCP schema.
   """
-  @spec encode_request(map(), String.t() | integer()) :: {:ok, String.t()} | {:error, Error.t()}
+  def validate_message(message) when is_map(message) do
+    with {:error, _} <- mcp_message_schema(message) do
+      {:error, :invalid_message}
+    end
+  end
+
+  @doc """
+  Encodes a request message to a JSON-RPC 2.0 compliant string.
+
+  Returns the encoded string with a newline character appended.
+  """
   def encode_request(request, id) do
-    data =
-      request
-      |> Map.put("jsonrpc", "2.0")
-      |> Map.put("id", id)
+    schema = get_schema(:request_schema)
 
-    {:ok, JSON.encode!(data) <> "\n"}
+    request
+    |> Map.put("jsonrpc", "2.0")
+    |> Map.put("id", id)
+    |> encode_message(schema)
   end
 
   @doc """
-  Encodes a notification message into a JSON-RPC 2.0 compliant string.
+  Encodes a notification message to a JSON-RPC 2.0 compliant string.
 
-  ## Parameters
-
-    * `notification` - A map containing the notification method and params
-
-  ## Returns
-
-    * `{:ok, json_string}` with the encoded notification and a newline
-    * `{:error, error}` if encoding fails
-
-  ## Examples
-
-      iex> {:ok, notif} = Hermes.MCP.Message.encode_notification(%{"method" => "notifications/progress", "params" => %{"progress" => 50}})
-      iex> notif_map = notif |> String.trim() |> JSON.decode!()
-      iex> notif_map["jsonrpc"]
-      "2.0"
-      iex> notif_map["method"]
-      "notifications/progress"
-      iex> notif_map["params"]["progress"]
-      50
+  Returns the encoded string with a newline character appended.
   """
-  @spec encode_notification(map()) :: {:ok, String.t()} | {:error, Error.t()}
   def encode_notification(notification) do
-    data = Map.put(notification, "jsonrpc", "2.0")
+    schema = get_schema(:notification_schema)
 
-    {:ok, JSON.encode!(data) <> "\n"}
+    notification
+    |> Map.put("jsonrpc", "2.0")
+    |> encode_message(schema)
   end
 
   @doc """
-  Encodes a progress notification message.
+  Encodes a progress notification message to a JSON-RPC 2.0 compliant string.
 
   ## Parameters
 
@@ -208,24 +234,10 @@ defmodule Hermes.MCP.Message do
     * `progress` - The current progress value (number)
     * `total` - Optional total value for the operation (number)
 
-  ## Returns
-
-    * `{:ok, json_string}` with the encoded notification and a newline
-    * `{:error, error}` if encoding fails
-
-  ## Examples
-
-      iex> {:ok, notif} = Hermes.MCP.Message.encode_progress_notification("token123", 50)
-      iex> notif_map = notif |> String.trim() |> JSON.decode!()
-      iex> notif_map["method"]
-      "notifications/progress"
-      iex> notif_map["params"]["progressToken"]
-      "token123"
-      iex> notif_map["params"]["progress"]
-      50
+  Returns the encoded string with a newline character appended.
   """
   @spec encode_progress_notification(String.t() | integer(), number(), number() | nil) ::
-          {:ok, String.t()} | {:error, Error.t()}
+          {:ok, String.t()} | {:error, term()}
   def encode_progress_notification(progress_token, progress, total \\ nil)
       when (is_binary(progress_token) or is_integer(progress_token)) and is_number(progress) do
     params = %{
@@ -242,119 +254,76 @@ defmodule Hermes.MCP.Message do
   end
 
   @doc """
-  Validates a message according to the MCP protocol.
+  Encodes a response message to a JSON-RPC 2.0 compliant string.
 
-  Performs basic validation to ensure the message conforms to the JSON-RPC 2.0
-  structure and has valid MCP-specific fields.
-
-  ## Parameters
-
-    * `message` - The parsed JSON message to validate
-
-  ## Returns
-
-    * `:ok` if the message is valid
-    * `{:error, error}` if validation fails
-
-  ## Examples
-
-      iex> Hermes.MCP.Message.validate(%{"jsonrpc" => "2.0", "result" => %{}, "id" => "req_123"})
-      :ok
-      
-      iex> {:error, error} = Hermes.MCP.Message.validate(%{"jsonrpc" => "1.0", "result" => %{}, "id" => "req_123"})
-      iex> error.reason
-      :invalid_request
+  Returns the encoded string with a newline character appended.
   """
-  @spec validate(map()) :: :ok | {:error, Error.t()}
-  def validate(message) when not is_map(message) do
-    {:error, Error.invalid_request(%{reason: "not_a_map"})}
-  end
+  def encode_response(response, id) do
+    schema = get_schema(:response_schema)
 
-  def validate(message) when is_map(message) and not is_map_key(message, "jsonrpc") do
-    {:error, Error.invalid_request(%{reason: "missing_jsonrpc_field"})}
+    response
+    |> Map.put("jsonrpc", "2.0")
+    |> Map.put("id", id)
+    |> encode_message(schema)
   end
-
-  def validate(%{"jsonrpc" => version}) when version != "2.0" do
-    {:error, Error.invalid_request(%{reason: "invalid_jsonrpc_version"})}
-  end
-
-  def validate(message) when is_request(message) do
-    validate_request(message)
-  end
-
-  def validate(message) when is_notification(message) do
-    validate_notification(message)
-  end
-
-  def validate(message) when is_response(message) do
-    :ok
-  end
-
-  def validate(message) when is_error(message) do
-    :ok
-  end
-
-  def validate(message) when is_map(message) do
-    {:error, Error.invalid_request(%{reason: "invalid_message_structure"})}
-  end
-
-  defp validate_request(%{"method" => method} = message) do
-    with :ok <- validate_method(method, @request_methods) do
-      validate_id(message["id"])
-    end
-  end
-
-  defp validate_notification(%{"method" => method}) do
-    validate_method(method, @notification_methods)
-  end
-
-  defp validate_method(method, allowed_methods) do
-    if method in allowed_methods do
-      :ok
-    else
-      {:error, Error.method_not_found(%{method: method})}
-    end
-  end
-
-  defp validate_id(id) when is_binary(id) or is_integer(id), do: :ok
-  defp validate_id(_), do: {:error, Error.invalid_request(%{reason: "invalid_id_type"})}
 
   @doc """
-  Converts a message to the appropriate domain object based on its type.
+  Encodes an error message to a JSON-RPC 2.0 compliant string.
+
+  Returns the encoded string with a newline character appended.
+  """
+  def encode_error(response, id) do
+    schema = get_schema(:error_schema)
+
+    response
+    |> Map.put("jsonrpc", "2.0")
+    |> Map.put("id", id)
+    |> encode_message(schema)
+  end
+
+  defp encode_message(data, schema) do
+    encoder = {schema, {:transform, fn data -> JSON.encode!(data) <> "\n" end}}
+    Peri.validate(encoder, data)
+  end
+
+  @doc """
+  Generates a unique progress token that can be used for tracking progress.
+
+  This is a convenience function to create tokens for the progress notification system.
+  Returns a string token with "progress_" prefix followed by a Base64-encoded unique identifier.
+  """
+  @spec generate_progress_token() :: String.t()
+  def generate_progress_token do
+    binary = <<
+      System.system_time(:nanosecond)::64,
+      :erlang.phash2({node(), self()}, 16_777_216)::24,
+      :erlang.unique_integer()::32
+    >>
+
+    "progress_" <> Base.url_encode64(binary)
+  end
+
+  @doc """
+  Encodes a log message notification to be sent to the client.
 
   ## Parameters
 
-    * `message` - A parsed JSON-RPC 2.0 message
+    * `level` - The log level (debug, info, notice, warning, error, critical, alert, emergency)
+    * `data` - The data to be logged (any JSON-serializable value)
+    * `logger` - Optional name of the logger issuing the message
 
-  ## Returns
-
-    * `{:ok, domain_object}` with the appropriate type (Response or Error)
-    * `{:error, error}` if conversion fails
-
-  ## Examples
-
-      iex> resp = %{"jsonrpc" => "2.0", "result" => %{"data" => "value"}, "id" => "req_123"}
-      iex> {:ok, response} = Hermes.MCP.Message.to_domain(resp)
-      iex> response.__struct__
-      Hermes.MCP.Response
-      
-      iex> err = %{"jsonrpc" => "2.0", "error" => %{"code" => -32700, "message" => "Parse error"}, "id" => "req_123"}
-      iex> {:ok, error} = Hermes.MCP.Message.to_domain(err)
-      iex> error.__struct__
-      Hermes.MCP.Error
+  Returns the encoded notification string with a newline character appended.
   """
-  @spec to_domain(message()) ::
-          {:ok, Response.t() | Error.t()}
-          | {:error, Error.t()}
-  def to_domain(message) when is_response(message) do
-    {:ok, Response.from_json_rpc(message)}
+  @spec encode_log_message(String.t(), term(), String.t() | nil) :: {:ok, String.t()} | {:error, term()}
+  def encode_log_message(level, data, logger \\ nil) when level in @log_levels do
+    params = maybe_add_logger(%{"level" => level, "data" => data}, logger)
+
+    encode_notification(%{
+      "method" => "notifications/message",
+      "params" => params
+    })
   end
 
-  def to_domain(message) when is_error(message) do
-    {:ok, Error.from_json_rpc(message["error"])}
-  end
-
-  def to_domain(_message) do
-    {:error, Error.invalid_request(%{reason: "cannot_convert_to_domain"})}
-  end
+  defp maybe_add_logger(params, nil), do: params
+  defp maybe_add_logger(params, logger) when is_binary(logger), do: Map.put(params, "logger", logger)
 end
